streamer/requirements.txt
requests==2.30.0
loguru==0.6.0
kafka-python==2.0.2
gql[all]